Strip pseudos from utility selectors before check; fixes #45
