+++ lib/is-valid-utility.js
@@ -1,6 +1,12 @@
 'use strict';
 
 /**
+ * Module dependencies
+ */
+
+var listSequences = require('./listSequences');
+
+/**
  * Module exports
  */
 
@@ -12,5 +18,8 @@ module.exports = isValidUtility;
  * @returns {Boolean}
  */
 function isValidUtility(selector, pattern) {
-  return pattern.test(selector);
+  var sequences = listSequences(selector);
+  return sequences.every(function (sequence) {
+    return pattern.test(sequence);
+  });
 }
