+++ lib/listSequences.js
@@ -1,0 +1,25 @@
+module.exports = listSequences;
+
+/**
+ * Extract an array of selector sequences from a selector string ---
+ * all the sequences that were combined via combinators.
+ *
+ * CSS combinators are whitespace, >, +, ~
+ * (cf. http://www.w3.org/TR/css3-selectors/#selector-syntax)
+ *
+ * Ignore pseudo-selectors ... by presuming they come at the end of the
+ * sequence and cutting them off from the string that gets checked.
+ *
+ * @param {String} selector
+ * @returns {String[]}
+ */
+function listSequences(selector) {
+  return selector
+    .split(/[\s>+~]/)
+    .filter(function (s) {
+      return s !== '';
+    })
+    .map(function (s) {
+      return s.split(':')[0];
+    });
+}
